複窓作成できるようにファイル分け 一応起動して画面を映す程度でバグがでないくらいになったが色々不具合あり
